Add files via upload
This attempts to give the birds a "curved" directional movement by dividing the birds into two separate pieces: a circle, which moves according to the flocking behavior, and a triangle, which moves relative to the circle and "chases" the direction of the bird. This amounted to doing a lot of algebra and creating functions to do this algebra at each time step. The algebra can be found as functions under the Bird class. There are still significant problems, including making the rotation appear "smooth"